+++ lib/deimos/schema_backends/avro_base.rb
@@ -15,7 +15,20 @@ module Deimos
       # @override
       def initialize(schema:, namespace:)
         super(schema: schema, namespace: namespace)
-        @schema_store = AvroTurf::MutableSchemaStore.new(path: Deimos.config.schema.path)
+        path = Deimos.config.schema.path.presence || Deimos.config.schema.paths[:avro].first
+        if path.blank?
+          raise "No schema paths configured for `avro` backend!"
+        end
+        @schema_store = AvroTurf::MutableSchemaStore.new(path: path)
+      end
+
+      def supports_key_schemas?
+        true
+      end
+
+      # @return [Boolean]
+      def supports_class_generation?
+        true
       end
 
       # @override